added HealthComponent and HitboxComponent, restructured folders
To create a new enemy/environmental object with a health and hitbox component: drag health_component.tscn and hitbox_component.tscn into the new character, add a collisionPolygon2D  as a child of the HitboxComponent, and set the HitboxComponent's Health Component (done so on the inspector) as the relevant HealthComponent.

Custom HP values can be set in the newly created HealthComponent's inspector.